finish JdbcMealRepo

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -40,13 +41,14 @@ public class JdbcMealRepository implements MealRepository {
 				.addValue("id", meal.getId())
 				.addValue("dateTime", meal.getDateTime())
 				.addValue("description", meal.getDescription())
-				.addValue("calories", meal.getCalories());
+				.addValue("calories", meal.getCalories())
+				.addValue("user_id", userId);
 
 		if (meal.isNew()) {
 			Number newKey = insertUser.executeAndReturnKey(map);
 			meal.setId(newKey.intValue());
 		} else if (namedParameterJdbcTemplate.update(
-				"UPDATE meals SET date_time=:dateTime, description=:description, calories=:calories " +
+				"UPDATE meals SET user_id=:user_id, date_time=:dateTime, description=:description, calories=:calories " +
 						"WHERE id=:id", map) == 0) {
 			return null;
 		}
@@ -55,18 +57,18 @@ public class JdbcMealRepository implements MealRepository {
 
 	@Override
 	public boolean delete(int id, int userId) {
-		return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
+		return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
 	}
 
 	@Override
 	public Meal get(int id, int userId) {
-		List<Meal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=?", ROW_MAPPER, id);
+		List<Meal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
 		return DataAccessUtils.singleResult(meals);
 	}
 
 	@Override
 	public List<Meal> getAll(int userId) {
-		return jdbcTemplate.query("SELECT * FROM meals ORDER BY date_time", ROW_MAPPER);
+		return jdbcTemplate.query("SELECT * FROM meals WHERE user_id=? ORDER BY date_time DESC", ROW_MAPPER, userId);
 	}
 
 	@Override
